Make typed code runable!
There was a quite glaring issue with the existing code. It could not be
executed at runtime! While mypy treats the `JSONSchema` type as it
should, `JSONSchema["a string"]` would fail at runtime because
`JSONSchema` is a subclass of dict. As such it is not subscriptable!

There was no point to making it a dict subclass to begin with, we
just need something that is a suitable no-op at runtime. It is now
simply an alias to `typing.Type`. The class should never be used at
runtime anyway, so this is not a regression.